tools/requirements.txt
# RAG Index Creator dependencies
torch
sentence-transformers
faiss-cpu
datasets
tqdm
numpy
